refactor: use N-API
