+++ binding.gyp
@@ -1,0 +1,18 @@
+{
+  "targets": [
+    {
+      "target_name": "addon",
+      "cflags!": [ "-fno-exceptions" ],
+      "cflags_cc!": [ "-fno-exceptions" ],
+      "conditions":[
+        ["OS=='win'", {
+      	  "sources": [ "./src/native/win32_swca/lib/addon.cpp" ]
+      	}]
+      ], 
+      "include_dirs": [
+        "<!@(node -p \"require('node-addon-api').include\")"
+      ],
+      'defines': [ 'NAPI_DISABLE_CPP_EXCEPTIONS' ],
+    }
+  ]
+}
