+++ src/native/win32_swca/lib/swca.h
@@ -1,0 +1,20 @@
+#include <dwmapi.h>
+
+#include <iostream>
+
+struct ACCENTPOLICY {
+    int nAccentState;
+    int nFlags;
+    int nColor;
+    int nAnimationId;
+};
+
+struct WINCOMATTRPDATA {
+    int nAttribute;
+    PVOID pData;
+    ULONG ulDataSize;
+};
+
+typedef BOOL(WINAPI* pSetWindowCompositionAttribute)(HWND, WINCOMATTRPDATA*);
+
+int swca(HWND hwnd, int accentState, int color);
